# server/spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "virtu API",
    "description": "SimpleLogin-compatible email alias API. Authenticate with the `Authentication: <api_key>` header (SimpleLogin's exact header name — not Authorization). Errors use the {\"error\": \"...\"} envelope.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authentication",
        "description": "API key obtained from POST /auth/login"
      }
    },
    "schemas": {
...
      "NewCustomAliasRequest": {
        "type": "object",
        "properties": {
          "alias_prefix": {
            "type": "string"
          },
          "signed_suffix": {
            "type": "string"
          },
          "mailbox_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991
            }
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "alias_prefix",
          "signed_suffix",
          "mailbox_ids"
        ],
        "example": {
          "alias_prefix": "groupon",
          "signed_suffix": ".abc12@virtu.email.XXXX.YYYY",
          "mailbox_ids": [
            1
          ],
          "note": "For groupon.com",
          "name": "Group On"
        }
      },
      "NewRandomAliasRequest": {
        "type": "object",
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "description": "SimpleLogin-compatible error envelope"
      },
...
      "MailboxLite": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email"
        ],
        "additionalProperties": false
      },
      "LatestActivity": {
        "type": "object",
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActivityAction"
          },
          "timestamp": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "contact": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string"
              },
              "name": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "reverse_alias": {
                "type": "string"
              }
            },
            "required": [
              "email",
              "name",
              "reverse_alias"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "action",
          "timestamp",
          "contact"
        ],
        "additionalProperties": false
      },
      "ActivityAction": {
        "type": "string",
        "enum": [
          "forward",
          "reply",
          "block",
          "bounced"
        ]
      },
...
      "AliasOptionsResponse": {
        "type": "object",
        "properties": {
          "can_create": {
            "type": "boolean"
          },
          "prefix_suggestion": {
            "type": "string"
          },
          "suffixes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AliasSuffix"
            }
          },
          "recommendation": {
            "description": "Present when ?hostname= is given and the user already created an alias for that hostname (latest one wins).",
            "type": "object",
            "properties": {
              "alias": {
                "type": "string"
              },
              "hostname": {
                "type": "string"
              }
            },
            "required": [
              "alias",
              "hostname"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "can_create",
          "prefix_suggestion",
          "suffixes"
        ],
        "additionalProperties": false
      },
      "AliasSuffix": {
        "type": "object",
        "properties": {
          "suffix": {
            "type": "string"
          },
          "signed_suffix": {
            "type": "string"
          },
          "is_custom": {
            "type": "boolean"
          },
          "is_premium": {
            "type": "boolean"
          }
        },
        "required": [
          "suffix",
          "signed_suffix",
          "is_custom",
          "is_premium"
        ],
        "additionalProperties": false
      },
      "CreatedAliasResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "email": {
            "type": "string"
          },
          "creation_date": {
            "type": "string"
          },
          "creation_timestamp": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "enabled": {
            "type": "boolean"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "nb_forward": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "nb_block": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "nb_reply": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "mailbox": {
            "$ref": "#/components/schemas/MailboxLite"
          },
          "mailboxes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MailboxLite"
            }
          },
          "support_pgp": {
            "type": "boolean"
          },
          "disable_pgp": {
            "type": "boolean"
          },
          "latest_activity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LatestActivity"
              },
              {
                "type": "null"
              }
            ]
          },
          "pinned": {
            "type": "boolean"
          },
          "alias": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "creation_date",
          "creation_timestamp",
          "enabled",
          "note",
          "name",
          "nb_forward",
          "nb_block",
          "nb_reply",
          "mailbox",
          "mailboxes",
          "support_pgp",
          "disable_pgp",
          "latest_activity",
          "pinned",
          "alias"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/v5/alias/options": {
      "get": {
        "tags": [
          "Alias"
        ],
        "description": "Options for creating a new alias: available (signed) suffixes, whether the user can create one, and a prefix suggestion derived from ?hostname=.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "hostname"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AliasOptionsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v3/alias/custom/new": {
      "post": {
        "tags": [
          "Alias"
        ],
        "description": "Create a new custom alias from a prefix and a signed suffix obtained from GET /v5/alias/options. The first mailbox_ids entry becomes the primary mailbox; the rest become extra delivery mailboxes. ?hostname= is recorded and drives the options `recommendation`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCustomAliasRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "hostname"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedAliasResponse"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "412": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/alias/random/new": {
      "post": {
        "tags": [
          "Alias"
        ],
        "description": "Create a new random alias on the user's default alias domain. ?mode=uuid|word overrides the user's alias_generator setting. ?hostname= is recorded and drives the options `recommendation`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "x-fastify-zod-openapi-optional": true,
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/NewRandomAliasRequest"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "hostname"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "mode"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedAliasResponse"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api",
      "description": "API root"
    }
  ],
...
}